refactor: move faucet creation to dbc genesis file
